Move compute_value to the json specific mod.
